vsvoice/requirements.txt
kivy==2.3.0
plyer==2.1.0
pyjnius==1.6.1